feat: split JWT / PAT authentication & make PAT auth work
