+++ packages/api-security/src/plugins/security.ts
@@ -1,11 +1,13 @@
 import {
+    GraphQLContext,
     GraphQLContextPlugin,
     GraphQLMiddlewarePlugin,
     GraphQLSchemaPlugin
 } from "@webiny/graphql/types";
 import { shield } from "graphql-shield";
-import authenticate from "./authentication/authenticate";
+import authenticateJwt from "./authentication/authenticateJwt";
 import { SecurityPlugin } from "@webiny/api-security/types";
+import LambdaClient from "aws-sdk/clients/lambda";
 
 const shieldMiddleware: GraphQLMiddlewarePlugin = {
     type: "graphql-middleware",
@@ -32,7 +34,7 @@ const shieldMiddleware: GraphQLMiddlewarePlugin = {
 
         return middleware;
     }
-} as GraphQLMiddlewarePlugin;
+};
 
 const middlewarePlugin = (options): GraphQLContextPlugin => ({
     type: "graphql-context",
@@ -54,8 +56,43 @@ const middlewarePlugin = (options): GraphQLContextPlugin => ({
     }
 });
 
+const authenticatePat = options => async (context: GraphQLContext) => {
+    if (context.user) {
+        return;
+    }
+
+    const { event } = context;
+    const { headers = {} } = event;
+    const authorization = headers["Authorization"] || headers["authorization"] || "";
+
+    if (!authorization) return;
+
+    const token = authorization;
+    const Lambda = new LambdaClient({ region: process.env.AWS_REGION });
+    const user = JSON.parse(
+        (
+            await Lambda.invoke({
+                FunctionName: options.validateAccessTokenFunction,
+                Payload: JSON.stringify({ PAT: token })
+            }).promise()
+        ).Payload as string
+    );
+
+    context.token = token;
+    context.user = user;
+};
+
 export default options => [
     shieldMiddleware,
     middlewarePlugin(options),
-    { type: "graphql-security", name: "graphql-security", authenticate }
+    {
+        type: "graphql-security",
+        name: "graphql-security-jwt",
+        authenticate: authenticateJwt
+    } as SecurityPlugin,
+    {
+        type: "graphql-security",
+        name: "graphql-security-pat",
+        authenticate: authenticatePat(options)
+    } as SecurityPlugin
 ];
